+++ src/commands/index.ts
@@ -5,3 +5,4 @@ export { createBoxPlotCommand } from "./boxplot";
 export { createTableCommand } from "./table";
 export { createEntitySummaryCommand, createEntitySummaryPerRunCommand } from "./entityBreakdown";
 export { createEntityMatrixCommand } from "./entityMatrix";
+export { createEntityHeatmapCommand } from "./entityHeatmap";